# api-contracts/openapi/openapi.yaml
openapi: 3.0.0
paths:
  /provision/onboard:
    post:
      operationId: ProvisioningController_onboard
      parameters: []
      requestBody:
        required: true
        description: Gateway onboard request with factory credentials and CSR
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardRequestDto'
      responses:
        '201':
          description: Gateway provisioned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardResponseDto'
        '400':
          description: Invalid request - malformed CSR or missing required fields
        '409':
          description: Conflict - gateway already provisioned
        '500':
          description: Server error - CA or NATS service unavailable
      tags:
        - Provisioning
info:
  title: NoTIP Provisioning Service
  description: NoTIP Provisioning Service OpenAPI specification
  version: 1.0.0
  contact: {}
tags: []
servers: []
components:
  schemas:
    FactoryCredentialsDto:
      type: object
      properties:
        factoryId:
          type: string
          description: Identificativo univoco del gateway (Opaque identifier)
          example: GW-NOTIP-2026-001
        factoryKey:
          type: string
          description: Segreto monouso pre-condiviso per l'autenticazione in fabbrica
          example: super-secret-factory-key
          writeOnly: true
      required:
        - factoryId
        - factoryKey
    OnboardRequestDto:
      type: object
      properties:
        credentials:
          description: Credenziali di fabbrica per autenticare il gateway
          allOf:
            - $ref: '#/components/schemas/FactoryCredentialsDto'
        csr:
          type: string
          description: Certificate Signing Request in formato PEM
          example: |-
            -----BEGIN CERTIFICATE REQUEST-----
            MIIB... 
            -----END CERTIFICATE REQUEST-----
        sendFrequencyMs:
          type: integer
          description: Frequenza di invio telemetria del gateway in millisecondi
          example: 5000
          minimum: 1
        firmwareVersion:
          type: string
          description: Versione firmware del gateway
          example: 1.0.0
      required:
        - credentials
        - csr
        - sendFrequencyMs
        - firmwareVersion
    GatewayIdentityDto:
      type: object
      properties:
        gatewayId:
          type: string
          description: Identificativo univoco del gateway
          example: gw-1
        tenantId:
          type: string
          description: Identificativo del tenant
          example: tenant-1
      required:
        - gatewayId
        - tenantId
    OnboardResponseDto:
      type: object
      properties:
        certPem:
          type: string
          description: Certificato foglia firmato dalla CA per il gateway (PEM)
          example: |-
            -----BEGIN CERTIFICATE-----
            MIID... 
            -----END CERTIFICATE-----
        aesKey:
          type: string
          description: Chiave AES-256 generata casualmente, codificata in Base64
          example: SGVsbG8gV29ybGQgQUVTLTI1NiBLZXk=
        identity:
          description: Identita del gateway validata dal servizio di management
          allOf:
            - $ref: '#/components/schemas/GatewayIdentityDto'
        sendFrequencyMs:
          type: integer
          description: Frequenza di invio telemetria del gateway in millisecondi
          example: 5000
          minimum: 1
      required:
        - certPem
        - aesKey
        - identity
        - sendFrequencyMs